authentification

+++ src/Controller/UserClientController.php
@@ -2,7 +2,7 @@
 
 namespace App\Controller;
 
-use App\Entity\User;
+use App\Entity\UserClient;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use Symfony\Component\HttpFoundation\Response;
@@ -11,64 +11,80 @@ use FOS\RestBundle\Controller\Annotations\View;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\Validator\ConstraintViolationList;
 
-class UserController extends AbstractFOSRestController
+class UserClientController extends AbstractFOSRestController
 {
     /**
-     * @Rest\Get("/users", name="app_user_list")
+     * @Rest\Get("api/users", name="app_user_list")
      * @View(serializerGroups={"list"})
      */
     public function getUsers()
     {
-        return $this->getDoctrine()->getRepository(User::class)->findAll();
+        $list = $this->getDoctrine()->getRepository(UserClient::class)->findBy(['user' => $this->getUser()]);
+
+        if (empty($list)) {
+            return $this->view(['message' => 'List Empty'], 200);
+        }
+
+        return $list;
     }
 
     /**
-     * @Rest\Get("/user/{id}", name="app_user_show", requirements={"id"="\d+"})
-     * @param User $user
+     * @Rest\Get("api/user/{id}", name="app_user_show", requirements={"id"="\d+"})
+     * @param UserClient $userClient
      * @View(serializerGroups={"show"})
      */
-    public function showUser(User $user)
+    public function showUser(UserClient $userClient)
     {
-        return $user;
+        if ($userClient->getUser() != $this->getUser()) {
+            return $this->view(['message' => 'Vous n\'etes pas le createur de cette utilisateur'], Response::HTTP_BAD_REQUEST);
+        }
+
+        return $userClient;
     }
 
     /**
-     * @Rest\Post("/user", name="app_user_create")
-     * @param User $user
+     * @Rest\Post("api/user", name="app_user_create")
+     * @param UserClient $userClient
      * @param ConstraintViolationList $violations
      * @View(statusCode=201,serializerGroups={"show"})
      * @ParamConverter(
-     *     "user",
+     *     "userClient",
      *     converter="fos_rest.request_body",
      *     options={
      *          "validator"={ "groups"="Create" }
      *     }
      * )
      */
-    public function createUser(User $user, ConstraintViolationList $violations)
+    public function createUser(UserClient $userClient, ConstraintViolationList $violations)
     {
         if (count($violations)) {
             return $this->view($violations, Response::HTTP_BAD_REQUEST);
         }
 
+        $userClient->setUser($this->getUser());
+
         $em = $this->getDoctrine()->getManager();
-        $em->persist($user);
+        $em->persist($userClient);
         $em->flush();
 
         return $this->view(
-            $user,
+            $userClient,
             201,
-            ['Location' => $this->generateUrl('app_user_show', ['id' => $user->getId(), UrlGeneratorInterface::ABSOLUTE_URL])]
+            ['Location' => $this->generateUrl('app_user_show', ['id' => $userClient->getId(), UrlGeneratorInterface::ABSOLUTE_URL])]
         );
     }
 
     /**
-     * @Rest\Delete("user/{id}", name="app_user_delete", requirements={"id"="\d+"})
-     * @param User $user
+     * @Rest\Delete("api/user/{id}", name="app_user_delete", requirements={"id"="\d+"})
+     * @param UserClient $user
      * @View(statusCode=204)
      */
-    public function deleteUser(User $user)
+    public function deleteUser(UserClient $user)
     {
+        if ($user->getUser() != $this->getUser()) {
+            return $this->view(['message' => 'Impossible'], Response::HTTP_CONFLICT);
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($user);
         $em->flush();
@@ -77,9 +93,9 @@ class UserController extends AbstractFOSRestController
     }
 
     /**
-     * @Rest\Put("user/{id}", name="app_user_update", requirements={"id"="\d+"})
-     * @param User $user
-     * @param User $userUpdate
+     * @Rest\Put("api/user/{id}", name="app_user_update", requirements={"id"="\d+"})
+     * @param UserClient $user
+     * @param UserClient $userUpdate
      * @param ConstraintViolationList $violations
      * @View(statusCode=200, serializerGroups={"show"})
      * @ParamConverter("user")
@@ -91,8 +107,12 @@ class UserController extends AbstractFOSRestController
      *     }
      * )
      */
-    public function updateUser(User $user, User $userUpdate, ConstraintViolationList $violations)
+    public function updateUser(UserClient $user, UserClient $userUpdate, ConstraintViolationList $violations)
     {
+        if ($user->getUser() != $this->getUser()) {
+            return $this->view(['message' => 'Vous n\'etes pas le createur de cette utilisateur'], Response::HTTP_BAD_REQUEST);
+        }
+
         if (count($violations)) {
             return $this->view($violations, Response::HTTP_BAD_REQUEST);
         }
